+++ src/components/organisms/teacher/OakUnitListOptionalityItem/OakUnitListOptionalityItem.stories.tsx
@@ -3,26 +3,54 @@ import { StoryObj, Meta } from "@storybook/react";
 
 import { OakUnitListOptionalityItem } from "./OakUnitListOptionalityItem";
 
-import { OakFlex } from "@/components/atoms";
+import { OakBox, OakFlex, OakHeading, OakP } from "@/components/atoms";
 
 const meta: Meta<typeof OakUnitListOptionalityItem> = {
   title: "components/organisms/teacher/OakUnitListOptionalityItem",
   component: OakUnitListOptionalityItem,
   tags: ["autodocs"],
   args: {
-    title:
-      "Migration: What do sources tell us about the British Empire in India and Africa?",
+    nullTitle: "British Empire in India and Africa",
     index: 1,
     href: "#",
     yearTitle: "Year 10",
     lessonCount: 10,
     isLegacy: false,
+    optionalityUnits: [
+      {
+        title:
+          "Migration: What do sources tell us about the British Empire in India and Africa?",
+
+        href: "#",
+        lessonCount: 10,
+      },
+      {
+        title: "Migration: What do sources tell us about the ",
+
+        href: "#",
+        lessonCount: 10,
+      },
+      {
+        title: "Migration: What do sources tell us ?",
+
+        href: "#",
+
+        lessonCount: 10,
+      },
+      {
+        title:
+          "Migration: What do sources tell us about the British Empire in India and Africa?",
+
+        href: "#",
+        lessonCount: 10,
+      },
+    ],
   },
   argTypes: {
     title: { control: { type: "text" } },
     yearTitle: { control: { type: "text" } },
     index: { control: { type: "number" } },
-    lessonCount: { control: { type: "number" } },
+    nullTitle: { control: { type: "text" } },
     disabled: { control: { type: "boolean" } },
     unavailable: { control: { type: "boolean" } },
     isLegacy: { control: { type: "boolean" } },
@@ -37,15 +65,39 @@ const meta: Meta<typeof OakUnitListOptionalityItem> = {
           $pa={"inner-padding-xl"}
           role="list"
         >
+          <OakBox>
+            <OakHeading tag="h3" $font={"heading-4"}>
+              Maths Unit
+            </OakHeading>
+            <OakP>
+              Brand-new teaching resources, thoughtfully crafted by teachers for
+              classroom needs.
+            </OakP>
+          </OakBox>
           {Story()}
-          {/* <OakUnitListOptionalityItem
-            title={"Numerals 1-10 (Legacy example)"}
-            lessonCount={10}
+          <OakUnitListOptionalityItem
+            nullTitle={"Numerals 1-10 (Legacy example) this is a long tite"}
             index={2}
             isLegacy={true}
             yearTitle="Year 10"
-            href={""}
-          /> */}
+            disabled={true}
+            optionalityUnits={[
+              {
+                title:
+                  "Migration: What do sources tell us about the British Empire in India and Africa?",
+
+                href: "#",
+                lessonCount: 10,
+              },
+              {
+                title:
+                  "Migration: What do sources tell us about the British Empire in India and Africa?",
+
+                href: "#",
+                lessonCount: 10,
+              },
+            ]}
+          />
         </OakFlex>
       );
     },
@@ -53,7 +105,7 @@ const meta: Meta<typeof OakUnitListOptionalityItem> = {
   parameters: {
     controls: {
       include: [
-        "title",
+        "nullTitle",
         "index",
         "lessonCount",
         "disabled",
