+++ src/components/organisms/teacher/OakUnitListOptionalityItem/OakUnitListOptionalityItem.tsx
@@ -3,80 +3,129 @@ import React, {
   ElementType,
   MutableRefObject,
 } from "react";
-import styled, { css } from "styled-components";
 
-import { OakFlex, OakSpan, OakHeading } from "@/components/atoms";
-import { parseColor } from "@/styles/helpers/parseColor";
-import { OakSecondaryLink } from "@/components/molecules";
-import { parseDropShadow } from "@/styles/helpers/parseDropShadow";
+import { OakUnitListOptionalityItemCard } from "../OakUnitListOptionalityItemCard";
 
-const StyledLessonLink = styled(OakSecondaryLink)`
-  text-decoration: none;
-`;
+import {
+  OakFlex,
+  OakBox,
+  OakHeading,
+  OakSpan,
+  OakLabel,
+  OakGrid,
+  OakGridArea,
+} from "@/components/atoms";
+import { FlexStyleProps } from "@/styles/utils/flexStyle";
 
-const StyledOptionalityListItem = styled(OakFlex)<{ $disabled?: boolean }>`
-  outline: none;
-  text-align: initial;
-  animation-timing-function: ease-out;
-  transition-duration: 300ms;
-
-  &:focus-visible {
-    background: ${parseColor("white")};
-    box-shadow: ${parseDropShadow("drop-shadow-centered-lemon")},
-      ${parseDropShadow("drop-shadow-centered-grey")};
-  }
-  ${(props) =>
-    !props.$disabled &&
-    css`  
-      cursor: pointer;
-
-      /* Don't apply hover styles on touch devices */
-      @media (hover: hover) {
-    &:hover {
-      ${StyledOakIndexBox} {
-        background: ${parseColor("lavender110")};
+const StyledYearAndOptionCount = ({
+  yearTitle,
+  optionalityUnitsLength,
+  disabledOrUnavailable,
+}: {
+  yearTitle: string | null | undefined;
+  optionalityUnitsLength: number;
+  disabledOrUnavailable: boolean | undefined;
+}) => (
+  <>
+    <OakSpan
+      $color={disabledOrUnavailable ? "text-disabled" : "text-primary"}
+      $mr={[null, "space-between-xxxl"]}
+    >
+      {yearTitle}
+    </OakSpan>
+    <OakBox
+      $ph={"inner-padding-xs"}
+      $ba={"border-solid-s"}
+      $borderColor={"bg-decorative4-main"}
+      $background={
+        disabledOrUnavailable
+          ? "bg-neutral-stronger"
+          : "bg-decorative4-very-subdued"
       }
-      background: ${parseColor("bg-decorative3-main")};
-    }
-  }
-      }
-    `}
+      $borderRadius={"border-radius-s"}
+      $ml={[null, "space-between-s"]}
+    >
+      <OakLabel
+        $color={disabledOrUnavailable ? "text-disabled" : "text-primary"}
+      >{`${optionalityUnitsLength} unit options`}</OakLabel>
+    </OakBox>
+  </>
+);
 
-  ${(props) =>
-    props.$disabled &&
-    css`
-      cursor: not-allowed;
-      background: ${parseColor("bg-neutral")};
-      border-color: ${parseColor("border-neutral-lighter")};
-    `}
-`;
+const StyledIndex = ({
+  index,
+  disabledOrUnavailable,
+  ...rest
+}: {
+  index: number;
+  disabledOrUnavailable: boolean | undefined;
+} & FlexStyleProps) => (
+  <OakFlex
+    $btlr={"border-radius-s"}
+    $bblr={"border-radius-s"}
+    $background={disabledOrUnavailable ? "bg-neutral-stronger" : "lavender"}
+    $minWidth={"all-spacing-11"}
+    $width={["all-spacing-11", "all-spacing-11", "auto"]}
+    $height={["all-spacing-11", "auto", "auto"]}
+    $justifyContent={"center"}
+    $alignItems={"center"}
+    {...rest}
+  >
+    <OakSpan
+      $font={"heading-5"}
+      $color={disabledOrUnavailable ? "text-disabled" : "text-primary"}
+    >
+      {index}
+    </OakSpan>
+  </OakFlex>
+);
 
-const StyledOakIndexBox = styled(OakFlex)`
-  border-top-left-radius: 6px;
-  border-bottom-left-radius: 6px;
-  animation-timing-function: ease-out;
-  transition-duration: 300ms;
-  @media (hover: hover) {
-    &:hover {
-      background: ${parseColor("lavender110")};
-    }
+const StyledUnitHeading = ({
+  nullTitle,
+  disabledOrUnavailable,
+}: {
+  nullTitle: string | undefined;
+  disabledOrUnavailable: boolean | undefined;
+}) => {
+  if (nullTitle === undefined) {
+    return null;
   }
-`;
+  return (
+    <OakFlex $maxWidth={"all-spacing-21"}>
+      <OakHeading
+        $color={disabledOrUnavailable ? "text-disabled" : "text-primary"}
+        $font={"heading-7"}
+        tag="h3"
+      >
+        {nullTitle}
+      </OakHeading>
+    </OakFlex>
+  );
+};
 
 export type OakUnitListOptionalityItemProps<C extends ElementType> = {
   as?: C;
   disabled?: boolean;
   unavailable?: boolean;
-  title: string;
-  lessonCount: number | null;
+  index: number;
+  nullTitle: string | undefined;
+  yearTitle?: string | null;
 
-  href: string;
-  firstItemRef?: MutableRefObject<HTMLAnchorElement | null> | null | undefined;
+  optionalityUnits: {
+    title: string;
+    href: string;
+
+    lessonCount: number;
+    firstItemRef?:
+      | MutableRefObject<HTMLAnchorElement | null>
+      | null
+      | undefined;
+  }[];
 } & ComponentPropsWithoutRef<C>;
 
 /**
  *
- * OakUnitsListItem component used as links for unit cards
+ * OakUnitsListOptionalityItem component used as links for unit cards with optionality
  */
 export const OakUnitListOptionalityItem = <C extends ElementType>(
   props: OakUnitListOptionalityItemProps<C>,
@@ -91,60 +140,94 @@ export const OakUnitListOptionalityItem = <C extends ElementType>(
     unavailable,
     onClick,
     index,
-
     firstItemRef,
+    optionalityUnits,
+    nullTitle,
     ...rest
   } = props;
 
   const disabledOrUnavailable = disabled || unavailable;
 
   return (
-    <OakFlex $boxSizing={"border-box"} $flexGrow={1} role="listitem">
-      <StyledOptionalityListItem
-        $pa={"inner-padding-m"}
-        $background={"bg-decorative3-very-subdued"}
-        $borderRadius="border-radius-m"
-        $borderColor={"border-decorative3"}
-        $ba="border-solid-m"
-        $disabled={disabledOrUnavailable}
-        href={disabledOrUnavailable ? undefined : href}
-        onClick={disabledOrUnavailable ? undefined : onClick}
-        ref={firstItemRef}
-        as={"a"}
-        $flexGrow={1}
-        {...rest}
-      >
+    <OakFlex
+      $flexDirection={["column", "row", "row"]}
+      $width={"100%"}
+      role="listitem"
+      {...rest}
+    >
+      <OakFlex $display={["flex", "none"]} $background={"white"}>
+        <StyledIndex
+          index={index}
+          disabledOrUnavailable={disabledOrUnavailable}
+        />
+
+        <OakFlex $alignItems={"center"} $ma={"space-between-xs"}>
+          <StyledUnitHeading
+            nullTitle={nullTitle}
+            disabledOrUnavailable={disabledOrUnavailable}
+          />
+        </OakFlex>
+      </OakFlex>
+      <StyledIndex
+        index={index}
+        disabledOrUnavailable={disabledOrUnavailable}
+        $display={["none", "flex", "flex"]}
+      />
+
+      <OakBox $background={"white"} $pa={"inner-padding-l"} $width={"100%"}>
         <OakFlex
-          $justifyContent={"space-between"}
-          $flexGrow={1}
-          $flexDirection={"column"}
+          $alignItems={"center"}
+          $mb={"space-between-s"}
+          $display={["none", "none", "flex"]}
         >
-          <OakHeading
-            $font={"heading-7"}
-            $color={disabledOrUnavailable ? "text-disabled" : "text-primary"}
-            tag={"h3"}
-            $mb={"space-between-xs"}
+          <OakFlex
+            $flexGrow={1}
+            $alignItems={"center"}
+            $justifyContent={"space-between"}
           >
-            {props.title}
-          </OakHeading>
-          <OakFlex $justifyContent={"flex-end"}>
-            <StyledLessonLink
-              isTrailingIcon
-              iconName="chevron-right"
-              disabled={disabledOrUnavailable}
-            >
-              <OakSpan
-                $font={"heading-light-7"}
-                $color={
-                  disabledOrUnavailable ? "text-disabled" : "text-primary"
-                }
-              >
-                {lessonCount} lessons
-              </OakSpan>
-            </StyledLessonLink>
+            <StyledUnitHeading
+              nullTitle={nullTitle}
+              disabledOrUnavailable={disabledOrUnavailable}
+            />
+            <OakFlex>
+              <StyledYearAndOptionCount
+                yearTitle={props.yearTitle}
+                optionalityUnitsLength={optionalityUnits.length}
+                disabledOrUnavailable={disabledOrUnavailable}
+              />
+            </OakFlex>
           </OakFlex>
         </OakFlex>
-      </StyledOptionalityListItem>
+        <OakBox $mb={"space-between-s"} $display={["none", "flex", "none"]}>
+          <StyledUnitHeading
+            nullTitle={nullTitle}
+            disabledOrUnavailable={disabledOrUnavailable}
+          />
+        </OakBox>
+        <OakFlex
+          $display={["flex", "flex", "none"]}
+          $mb={"space-between-s"}
+          $justifyContent={"space-between"}
+        >
+          <StyledYearAndOptionCount
+            yearTitle={props.yearTitle}
+            optionalityUnitsLength={optionalityUnits.length}
+            disabledOrUnavailable={disabledOrUnavailable}
+          />
+        </OakFlex>
+        <OakGrid $gridGap={"space-between-xs"} role="list">
+          {optionalityUnits.map((unit, index) => (
+            <OakGridArea key={`${unit.title}-${index}`} $colSpan={[12, 6]}>
+              <OakUnitListOptionalityItemCard
+                key={unit.title}
+                {...unit}
+                disabled={disabledOrUnavailable}
+                ref={index === 1 ? firstItemRef : null}
+              />
+            </OakGridArea>
+          ))}
+        </OakGrid>
+      </OakBox>
     </OakFlex>
   );
 };
