--- src/components/organisms/teacher/OakUnitListOptionalityUnit/index.ts
@@ -1,1 +1,0 @@
-export * from "./OakUnitListOptionalityUnit";
